rename kaltura-logger
