+++ kaltura-logger/.gitignore
