+++ kaltura-logger/src/index.ts
@@ -1,2 +1,2 @@
-export * from './kaltura-log.module';
+export * from './kaltura-logger.module';
 export * from './kaltura-logger.service';
