+++ kaltura-logger/src/kaltura-logger.service.ts
