+++ kaltura-logger/tsconfig.json
