+++ lerna.json
@@ -1,7 +1,7 @@
 {
   "lerna": "0.0.2",
   "packages": [
-    "kaltura-log",
+    "kaltura-logger",
     "kaltura-common",
     "kaltura-ui",
     "kaltura-primeng-ui",
